requirements.txt
torch
numpy
pyyaml
openpoints
